Changed scraping from datamine to scraping from pokeapi.

+++ src/modifiers.py
@@ -1,5 +1,4 @@
 import math
-from decimal import Decimal
 
 from models.exceptions import InvalidCatchingPowerError
 from models.poke_ball import PokeBall
@@ -28,7 +27,7 @@ def get_dark_grass_modifier(registered_pokemon_on_dex: int) -> int:
 
 
 def get_species_modifier(
-    catch_rate: int, weight: Decimal, poke_ball: PokeBall
+    catch_rate: int, weight: float, poke_ball: PokeBall
 ) -> int:
     if poke_ball == PokeBall.HEAVY_BALL:
         for (min_weight, max_weight), modifier in HEAVY_BALL_RANGES.items():
